requirements.txt
gspread
pandas
python-dotenv
tweepy
